refactor: 요일별 연재 리팩토링
