+++ package.json
@@ -15,7 +15,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.23.0",
-    "styled-components": "^6.1.11",
     "swiper": "^11.1.3",
     "vite-tsconfig-paths": "^4.3.2"
   },
