+++ src/pages/home/HomeStyle.ts
@@ -2,58 +2,58 @@ export const ARTICLES = [
   {
     id: 1,
     title: '나의 우울조각을 찾을 수 왔수다',
-    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는 ...',
+    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는홀로 선다는 건 한 번에 한 가지 일을 하는',
     author: '레몬숲',
-    imageSrc: '',
+    imageSrc: 'https://via.placeholder.com/56x58',
   },
   {
     id: 2,
     title: '나의 우울조각을 찾을 수 왔수다',
-    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는 ...',
+    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는',
     author: '레몬숲',
-    imageSrc: '',
+    imageSrc: 'https://via.placeholder.com/56x58',
   },
   {
     id: 3,
     title: '나의 우울조각을 찾을 수 왔수다',
-    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는 ...',
+    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는',
     author: '레몬숲',
-    imageSrc: '',
+    imageSrc: 'https://via.placeholder.com/56x58',
   },
   {
     id: 4,
     title: '나의 우울조각을 찾을 수 왔수다',
-    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는 ...',
+    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는',
     author: '레몬숲',
-    imageSrc: '',
+    imageSrc: 'https://via.placeholder.com/56x58',
   },
 
   {
     id: 5,
     title: '나의 우울조각을 찾을 수 왔수다',
-    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는 ...',
+    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는',
     author: '레몬숲',
-    imageSrc: '',
+    imageSrc: 'https://via.placeholder.com/56x58',
   },
   {
     id: 6,
     title: '나의 우울조각을 찾을 수 왔수다',
-    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는 ...',
+    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는',
     author: '레몬숲',
-    imageSrc: '',
+    imageSrc: 'https://via.placeholder.com/56x58',
   },
   {
     id: 7,
     title: '나의 우울조각을 찾을 수 왔수다',
-    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는 ...',
+    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는',
     author: '레몬숲',
-    imageSrc: '',
+    imageSrc: 'https://via.placeholder.com/56x58',
   },
   {
     id: 8,
     title: '나의 우울조각을 찾을 수 왔수다',
-    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는 ...',
+    subtitle: '홀로 선다는 건 한 번에 한 가지 일을 하는',
     author: '레몬숲',
-    imageSrc: '',
+    imageSrc: 'https://via.placeholder.com/56x58',
   },
 ];
